Struct oak_attestation_common::certificate::CertificateAuthority

pub struct CertificateAuthority {
    pub key_pair: PKey<Private>,
    pub root_certificate: X509,
}

Convenience structure for generating X.509 certificates. https://tools.ietf.org/html/rfc5280

Fields
key_pair: PKey<Private>root_certificate: X509
Implementations

impl CertificateAuthority

pub fn create(add_tee_extension: AddTeeExtension) -> Result<Self>

Generates a root X.509 certificate and a corresponding private/public key pair. add_tee_extension indicates whether to add a custom extension containing a TEE report to the root certificate.

pub fn sign_certificate(
&self,
request: X509Req,
add_tee_extension: AddTeeExtension
) -> Result<X509>

Generates an X.509 certificate based on the certificate signing request. add_tee_extension indicates whether to add a custom extension containing a TEE report.

pub fn get_private_key_pem(&self) -> Result<Vec<u8>>

Get RSA key pair encoded in PEM format. https://tools.ietf.org/html/rfc7468

pub fn get_root_certificate_pem(&self) -> Result<Vec<u8>>

Get a root X.509 certificate encoded in PEM format. https://tools.ietf.org/html/rfc7468